[py] [BinaryAccel] Always expect unicode strings.
`fastbinary` currently ignores the `utf8strings` flag. That is undesirable
because when `utf8strings` is set, `TBinaryProtocolAccelerated` will just
start failing serialization attempts of unicode strings.

This changes it to always expect UTF-8 strings.

This is a quick short-term fix. A more complete fix is to change the compiler
to pass whether `utf8strings` is set to the library and transform strings
accordingly. See fbthrift's fastbinary.c.