[python] fastbinary `utf8strings` support
`fastbinary` does not currently respect the `utf8strings` flag.
Relevant JIRA ticket: https://issues.apache.org/jira/browse/THRIFT-1229

This is an as yet unmerged patch submitted to the original JIRA ticket by
James Haggerty. It changes the Python code generator to include whether a type
is unicode or not in the `thrift_spec`, and `fastbinary` to use that
information to UTF-8 encode/decode values.

This change is *not compatible with Python 3*. The behavior will have to be
altered slightly to account for the changes in `str`/`bytes`/`unicode` in
Python 3.